+++ src/Constants/IdStrings.php
@@ -48,6 +48,7 @@ class IdStrings
     public const INVOICE_GUARANTEED = 'ivg';
     public const INVOICE_SECURED = 'ivs';
     public const KLARNA = 'kla';
+    public const PAYLATER_DIRECT_DEBIT = 'pdd';
     public const PAYLATER_INVOICE = 'piv';
     public const PAYLATER_INSTALLMENT = 'pit';
     public const PAYMENT_PAGE = 'ppg';
@@ -86,7 +87,9 @@ class IdStrings
         self::INVOICE_GUARANTEED,
         self::INVOICE_SECURED,
         self::KLARNA,
+        self::PAYLATER_DIRECT_DEBIT,
         self::PAYLATER_INVOICE,
+        self::PAYLATER_INSTALLMENT,
         self::PAYMENT_PAGE,
         self::PAYPAL,
         self::PAYU,