+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres
 to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.4.0](https://github.com/unzerdev/php-sdk/compare/3.3.0..3.4.0)
+### Added
+*   Add `\UnzerSDK\Resources\PaymentTypes\PaylaterDirectDebit` payment type.
+
+### Changed
+*   Allow 'null' for setters in '\UnzerSDK\Traits\HasAccountInformation' trait to avoid error when e.g. API response contains empty bic.
+
+### Deprecated
+*   `\UnzerSDK\Resources\PaymentTypes\SepaDirectDebit`, please use `\UnzerSDK\Resources\PaymentTypes\PaylaterDirectDebit'`.
+*   `\UnzerSDK\Resources\PaymentTypes\SepaDirectDebitSecured`, please use `\UnzerSDK\Resources\PaymentTypes\PaylaterDirectDebit'`.
+
 ## [3.3.0](https://github.com/unzerdev/php-sdk/compare/3.2.0..3.3.0)
 ### Added
 *   Chargeback transaction type.